part4/back/part3/requirement.txt
flask
flask-restx
flask-bcrypt
flask-jwt-extended
flask-sqlalchemy
flask-cors
python-dotenv
